moved service tab, admin tab removed
