+++ nussknacker-dist/src/universal/conf/dev-application.conf
@@ -248,7 +248,7 @@ environmentAlert: {
 
 tabs = ${tabs} [
   {title: "Metrics", type: "IFrame", url: ${grafanaUrl}"/dashboard/db/"${metricsSettings.defaultDashboard}"?theme=dark&var-env="${environment}, id: "metrics"},
-  {title: "Admin", type: "Local", url: "/admin", id: "admin", requiredPermission: "AdminTab"}
+  {title: "Services", type: "Local", url: "/services", id: "services", requiredPermission: "AdminTab"}
 ]
 
 componentLinks: [
