+++ ui/client/containers/paths.ts
@@ -8,7 +8,7 @@ export const VisualizationPath = `${VisualizationBasePath}/:processId`
 export const MetricsBasePath = `/metrics`
 export const MetricsPath = `${MetricsBasePath}/:processId?`
 export const SignalsPath = `/signals`
-export const AdminPagePath = `/admin`
+export const ServicesPath = `/services`
 export const CustomTabBasePath = `/customtabs`
 export const ScenariosBasePath = `/scenarios`
